- replaced the previous way of customization with new one wich uses ApplicationConfigurationPopulator
