+++ Server/pom.xml
@@ -31,6 +31,7 @@
 				<includes>
 					<include>**/*.xml</include>
 					<include>**/*.properties</include>
+					<include>META-INF/services/*.*</include>
 				</includes>
 			</resource>
 		</resources>
