+++ Server/src/main/resources/META-INF/navigation/super-gluu.navigation.xml
