+++ Server/src/main/resources/META-INF/services/javax.faces.application.ApplicationConfigurationPopulator
@@ -1,0 +1,1 @@
+org.xdi.oxauth.customization.FacesConfigPopulator
