+++ Server/src/main/webapp/WEB-INF/faces-config.xml
